Move .core to .reikai
