+++ core/src/main/kotlin/io/github/kerelape/reikai/EntityWrap.kt
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package io.github.kerelape.reikai.core
+package io.github.kerelape.reikai
 
 /**
  * Wrap envelope for [Entity].
