+++ core/src/main/kotlin/io/github/kerelape/reikai/io/Content.kt
@@ -23,8 +23,8 @@
  */
 package io.github.kerelape.reikai.io
 
-import io.github.kerelape.reikai.core.Entity
-import io.github.kerelape.reikai.core.EntityWrap
+import io.github.kerelape.reikai.Entity
+import io.github.kerelape.reikai.EntityWrap
 
 /**
  * All content of a [RandomAccessSource].
