+++ core/src/main/kotlin/io/github/kerelape/reikai/io/Destination.kt
@@ -23,7 +23,7 @@
  */
 package io.github.kerelape.reikai.io
 
-import io.github.kerelape.reikai.core.Entity
+import io.github.kerelape.reikai.Entity
 
 /**
  * Destination of data.
