+++ core/src/main/kotlin/io/github/kerelape/reikai/io/RandomAccessDestination.kt
@@ -23,7 +23,7 @@
  */
 package io.github.kerelape.reikai.io
 
-import io.github.kerelape.reikai.core.Entity
+import io.github.kerelape.reikai.Entity
 
 /**
  * Destination with random access.
