+++ core/src/main/kotlin/io/github/kerelape/reikai/io/Row.kt
@@ -23,8 +23,8 @@
  */
 package io.github.kerelape.reikai.io
 
-import io.github.kerelape.reikai.core.Entity
-import io.github.kerelape.reikai.core.Length
+import io.github.kerelape.reikai.Entity
+import io.github.kerelape.reikai.Length
 
 /**
  * Sequence of data.
