+++ core/src/main/kotlin/io/github/kerelape/reikai/io/Source.kt
@@ -23,7 +23,7 @@
  */
 package io.github.kerelape.reikai.io
 
-import io.github.kerelape.reikai.core.Entity
+import io.github.kerelape.reikai.Entity
 
 /**
  * Source of data.
