+++ core/src/main/kotlin/io/github/kerelape/reikai/io/filesystem/RandomAccessFileChannel.kt
@@ -23,8 +23,8 @@
  */
 package io.github.kerelape.reikai.io.filesystem
 
-import io.github.kerelape.reikai.core.Data
-import io.github.kerelape.reikai.core.Entity
+import io.github.kerelape.reikai.Data
+import io.github.kerelape.reikai.Entity
 import io.github.kerelape.reikai.io.RandomAccessChannel
 import io.github.kerelape.reikai.io.RandomAccessSource
 import io.github.kerelape.reikai.math.asEntity
