+++ core/src/main/kotlin/io/github/kerelape/reikai/io/network/TcpEndpoint.kt
@@ -23,7 +23,7 @@
  */
 package io.github.kerelape.reikai.io.network
 
-import io.github.kerelape.reikai.core.Entity
+import io.github.kerelape.reikai.Entity
 import io.github.kerelape.reikai.io.Channel
 import io.github.kerelape.reikai.io.Source
 import io.github.kerelape.reikai.logic.False
