+++ core/src/main/kotlin/io/github/kerelape/reikai/logic/Conversions.kt
@@ -23,7 +23,7 @@
  */
 package io.github.kerelape.reikai.logic
 
-import io.github.kerelape.reikai.core.Entity
+import io.github.kerelape.reikai.Entity
 import java.math.BigInteger
 
 /**
