+++ core/src/test/kotlin/io/github/kerelape/reikai/StructGuardTest.kt
@@ -1,5 +1,6 @@
-package io.github.kerelape.reikai.core
+package io.github.kerelape.reikai
 
+import io.github.kerelape.reikai.*
 import io.github.kerelape.reikai.logic.True
 import io.github.kerelape.reikai.math.asEntity
 import io.github.kerelape.reikai.text.Text
