+++ core/src/test/kotlin/io/github/kerelape/reikai/logic/BooleanGuardTest.kt
@@ -23,9 +23,9 @@
  */
 package io.github.kerelape.reikai.logic
 
-import io.github.kerelape.reikai.core.Data
-import io.github.kerelape.reikai.core.DataizationException
-import io.github.kerelape.reikai.core.Empty
+import io.github.kerelape.reikai.Data
+import io.github.kerelape.reikai.DataizationException
+import io.github.kerelape.reikai.Empty
 import io.github.kerelape.reikai.text.Text
 import kotlinx.coroutines.runBlocking
 import org.junit.jupiter.api.Assertions
