+++ core/src/test/kotlin/io/github/kerelape/reikai/logic/IdentityTest.kt
@@ -23,7 +23,7 @@
  */
 package io.github.kerelape.reikai.logic
 
-import io.github.kerelape.reikai.core.Empty
+import io.github.kerelape.reikai.Empty
 import io.github.kerelape.reikai.math.asEntity
 import kotlinx.coroutines.runBlocking
 import org.junit.jupiter.api.Assertions
